Combine uniforms into variables
