+++ packages/shader-composer/src/stdlib/index.ts
@@ -10,7 +10,6 @@ export * from "./rotation"
 export * from "./scene"
 export * from "./spaces"
 export * from "./textures"
-export * from "./uniforms"
 export * from "./values"
 export * from "./variables"
 export * from "./vectors"
